+++ cmd/xswap/session.go
@@ -20,6 +20,7 @@ var sessionIDPattern = regexp.MustCompile(`^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-
 
 type codexSession struct {
 	ID      string
+	Account string
 	CWD     string
 	Path    string
 	Preview string